Remove http_client function
There is no need to have a separate http_client function that only
simply wraps around Application.get_env called from a single place. In
the tests we are mocking HTTPoison by replacing the
configuration (environment) value, not by replacing this function. The
mocking this continues to work even without a dedicated function.

This also removes the need of having a variable of a different name
that would store the HTTP client module.