Replace first_url with input_url
Although @url[0] technically is the first URL in the list, it is not
relevant for its usage. There in the single URL test we need just any
URL and picking the first is a natural choice.

In contrast, the last_url in the multiple URLs test is semantic. The
output file is overwritten by every download, so only the last URL
remains there at the end. It is the only thing we can test.