Convert load_urls to a Stream
The downloading should work as a stream, everything from picking the
URL from the file through the download from the Internet to the local
drive file write should happen separately for every item, not by
repeated enumeration over a list. This required another version of the
download/3 function accepting a Stream and returning a Stream.