Remove download/3 with urls as list
Enumerable traversal is a trivial concept. It does not make sense to
have a separate function just to do Enum.map. It brings confusing
function dichotomy and a need for guard clauses. Also it discourages
from using other types of enumerables like streams.

For the function to be testable, the get_env call needed to be moved
to main/0, which makes nice symmetry with http_client. Because the
input_path parameter is of the same type and thus indistiguishable
from a URL, it only makes sense to rename the function.