Move get input path to main
It is not possible to pass a custom input file to the download
function and thus the function is not testable. The existing test in
fact does not test the file read at all. Fixed by moving the get_env
call to the main method, which makes a nice symmetry with
http_client.

Now, the test really tests what it should, the
configuration loads are collected to a single place and the functions
are even more atomic.

The input path is a string just as a URL and thus these two are
indistinguishable. Renamed the function to download_urls and
download_url to avoid this. Still using the ugly guards to distinguish
between the URL list and an input file path. Addressing that would be
over the scope of this changeset.