Remove a wrong test
The “download URLs from the input file” test actually did not test
anything else than the “download multiple URLs” test. Although a URLs
input file is created, it is not passed to the download function and
not used at all.

It is currently not possible to test this, because the download
function does not accept an input file as an argument. Removing the
test until that is fixed.