Rename download_urls that accepts a file
The two download_urls functions do different things. Used a longer
name for the one that accepts a file to communicate that it actually does
two things – loads a file and download URLs – and that it may be worth
further refactoring. This may disappear naturally when leveraging
Streams over lists.