Move get input path to load_urls
There is no practical need to have the URLs input path to be passed to
the function as an artgument. There is only going to be a single URL
input file and the tests too can load the value from the
environment.

The directory path still needs to be passed around because the
executive code uses the current working directory, whilst the tests
use a temporary directory.