Rename materialization to materializer
Initially, I considered the materialization a virtual step separate from Onigumo and the Spider. It, however, makes sense to enable Spiders to define ways to materialize the structured data as a natural step invokable from Onigumo.

That way, the materializer becomes a part of the Spider that does not have its Onigumo counterpart. Similarly, the Downloader exists only in Onigumo, but does not touch the Spider.