Extract common parts outside the loop
The lines that build the input path don’t need to reside inside the loop. The input_path_tmp value is the same for all iterations.

That shows the flaw of having the for loop inside the test. The temporary directory contains artifacts from the first run. It is just luck that it does not interfere with the second one.