Move for loops outside the tests
A loop inside the tests makes all the iterations run in the same test case. That means that if one of them fails, the subsequent ones don’t even run. They also share the environment, which causes no problem here, but as a practice, it can interfere later, e. g. with tests using temporary directories.