Use Map instead of Module.safe_concat
The CLI argument doesn’t need to be tightly coupled to the implementation detail of components being Elixir modules. ArgumentError raised by Module.safe_concat was only a symptom of that. Replaced with a mapping of string arguments (component names) to Elixir modules. As a result MatchError is raised when unknown component is passed, instead of a rather cryptic ArgumentError from Module.safe_concat.