Mock Onigumo.Downloader in CLI tests
Onigumo.Component is a behavior that can be mocked. CLI tests don’t
verify behavior of concrete components. Defined a mock for the
Downlaoader components so the CLI tests only verifies its call, not
its behavior.

Defined the expect mock to return the passed working directory to test
it’s set properly.