Update description to match diagrams
The Operator-Downloader-Parser trinity became only a part of the
whole architecture. It’s no longer true that Onigumo is composed of
only three components. It wasn’t entirely true before either: there
was still the idea of a Materializer.

There are two flowcharts now. It’d be possible to call them a single
one consisting of two parts, but clarity doesn’t harm.