Replace with with case
In the end, I found the `case` clause more expressive. There are three branches:

* Valid input: launch a component
* Invalid input: unknown switch
* Invalid input: wrong argument count

Having other scenarios in the future, there will be multiple shapes of
valid input. E.g. the -h/--help switch would introduce a branch with a
sole switch and no arguments.

The `with` clause seems to be more fitting for cases where we need to
collect input from multiple sources. That is partially the case here,
because we’re fetching from `@components`. Unfortunately however, I
don’t see a way to add that as a `case` branch.