Add a help switch and tests
This patch makes it possible to test the  change in the `pop`
branch. It adds a new `--help` switch and a `-h` alias. Only `-C` is
compatible with a componanet name. `-h dowloader` is thus an invalid
combination and should print the usage message. Without `-h`, there is
no other valid switch. Without one, there is no way to test the
behavior.